requirements.txt
fastapi
uvicorn
gitpython
requests
python-dotenv
openai
pydantic
